requirements.txt
sentence-transformers
requests
wikipedia-api
faiss-cpu
rank-bm25
transformers
torch
nltk
beautifulsoup4
requests
pandas
numpy
streamlit
tqdm
scikit-learn
matplotlib
accelerate
